Make AddEventWatch and friends work with go 1.6
1.6 introduces some extra restrictions on passing pointers into unsafe C
code, notably here we can't pass a pointer to an thing that contains a
pointer to a go object. It's a little silly in this case since the
pointer is completely opaque from SDL's perspective, but I didn't make
the restriction so don't shoot me.

We do this by generating unique handles for each event instead of
passing a pointer to the go structs. Both of these are unique, but one
doesn't make go upset.

I also renamed eventWatchesCache to eventWatches since it's not really a
cache anymore that just tracks the lifetime of unsafe objects(?) and is
the proper source of truth for events.