sdl: hints: Add initial implementation of AddHintCallback and DelHintCallback
This implementation use a Go map to store the Go callback functions mapped to
their hint names. There is a single C function used as the callback for all
native SDL_AddHintCallback and DelHintCallback. The C function calls an exported
Go function that calls the function in the map mapped to the hint name.

Signed-off-by: Lilis Iskandar <lilis@veand.co>