sdl: Fix computation of slice returned by LockTexture.
When SDL_LockTexture is given a rect the pointer that is returned points
to the very first pixel (X:0, Y:0 of the Rect) and contains a valid
memory address for every sub pixel coordinate after that until the final
pixel. This means that a 200x200 square inside of a 400x400 texture
should have a length of 319,200. The pitch and dimensions of the target
locked area should be used to skip over the portions that are unused.

Before this change the slice returned only had a length of the rectangle
dimensions and made it impossible to fill in the entire locked
rectangle.

The algorithm used to compute the length is as follows:
Compute the amount of bytes per pixel by dividing the pitch by the pixel
width. Then compute the number of pixels as though we were using the
entirety of each row in the rectangle. Then subtract out the the unused
leading pixels. Then subtract out the unused trailing pixels.