sdl/render: fix RenderGeometryRaw() to work on SDL2 2.0.18 and other versions such as SDL2 2.0.20
Because the API for SDL_RenderGeometryRaw() changed after it was first introduced in SDL2 2.0.18,
users using SDL2 2.0.20 or newer could not build the binding. Therefore, I made it so it would
work on both SDL2 2.0.20 and older versions of SDL2 by providing compatibility function that
wraps the underlying SDL_RenderGeometryRaw() function.

Signed-off-by: Lilis Iskandar <lilis@veand.co>