Avoid unnecessary string alloc and fix potential memory leak.
I am not sure if the code is correctly freeing the string that was
allocated or whether by setting '_device = nil' then a different
pointer is freed.

In any case, the code allocates a string that may not be needed if
'device == ""', which can be avoided.

This patch avoids the unnecessary allocation if 'device == ""' and it
uses the same allocated pointer in free (just in case this is
required).

See issue https://github.com/veandco/go-sdl2/issues/582.